+++ src/service/UsuarioService.ts
@@ -7,29 +7,32 @@ export class UsuarioService {
     pessoaService: PessoaService = new PessoaService();
 
     async criarUsuario(usuarioData: any): Promise<UsuarioEntity> {
-        const { idPesssoa, senha, } = usuarioData;
-        if (typeof idPesssoa !== 'number' || typeof senha !== 'string') {
-            throw new Error("Dados incorretos, verificar se é 'string' ou 'number'");
+        const { idPessoa, senha } = usuarioData;
+        if (typeof idPessoa !== 'number' || typeof senha !== 'string') {
+            throw new Error("Dados incorretos: Insira os dados em seus devidos formatos");
         }
-        if (!idPesssoa || !senha) {
-            throw new Error("É necesario inserir um id e um pessoa");
+        if (!idPessoa || !senha) {
+            throw new Error("Dados incompletos: Verifique se todos os campos foram preechidos");
         }
-        const verificarUsuario = await this.usuarioRepository.filterById(usuarioData);
-        if (verificarUsuario) {
+        const verificarUsuario = await this.usuarioRepository.filterByIdPessoa(idPessoa);
+        if (verificarUsuario.length > 0) {
             throw new Error("Esse usuario já esxite");
         }
-        const verificaPessoa = await this.pessoaService.listarPessoaPorId(idPesssoa);
+        const verificaPessoa = await this.pessoaService.listarPessoaPorId(idPessoa);
         if (!verificaPessoa) {
-            throw new Error("Não existe uma pessoa com esse id");
+            throw new Error("Está pessoa não está cadastrada");
         }
-        const novaUsuario = this.usuarioRepository.insertUsuario(new UsuarioEntity(undefined, idPesssoa, senha));
-        console.log("Service - Insert ", novaUsuario);
-        return novaUsuario;
+        const novoUsuario = this.usuarioRepository.insertUsuario(new UsuarioEntity(undefined, idPessoa, senha));
+        console.log("Service - Insert ", novoUsuario);
+        return novoUsuario;
     }
 
-    async listarUsuarioPorId(usuarioData: any): Promise<UsuarioEntity> {
+    async listarUsuarioPorId(usuarioData: any): Promise<UsuarioEntity[]> {
         const id: number = usuarioData;
         const usuario = await this.usuarioRepository.filterById(usuarioData);
+        if (usuario.length === 0) {
+            throw new Error("Usuario não cadastrado");
+        }
         console.log("Service - Filter ID ", usuario);
         return usuario;
     }
@@ -41,20 +44,24 @@ export class UsuarioService {
     }
 
     async atualizarUsuario(usuarioData: any): Promise<UsuarioEntity> {
-        const { id, idPessoa, senha } = usuarioData;
+        const { idPessoa, senha, id } = usuarioData;
         const verificar = await this.listarUsuarioPorId(id);
         if (!verificar) {
-            throw new Error("Usuario não localizada");
+            throw new Error("Usuario não localizado");
         }
-        const usuario = new UsuarioEntity(id, senha)
+        const usuario = new UsuarioEntity(id, idPessoa, senha)
         await this.usuarioRepository.updateSenha(usuarioData);
         console.log("Service - Update ", usuario);
         return usuario;
     }
 
     async deletarUsuario(usuarioData: any): Promise<UsuarioEntity> {
-        const { id, idPessoa, senha } = usuarioData;
-        const usuario = new UsuarioEntity(id, idPessoa, senha)
+        const { id, idPessoa } = usuarioData;
+        const usuario = new UsuarioEntity(id, idPessoa);
+        const verificarUsuario = await this.listarUsuarioPorId(id);
+        if (verificarUsuario.length === 0) {
+            throw new Error("Usuario não localizado");
+        }
         await this.usuarioRepository.deleteUsuario(usuarioData);
         console.log("Service - Delete ", usuario);
         return usuario;
